remove Field component and rename folders
