--- src/components/Form/Field.scss
@@ -1,40 +1,0 @@
-@use "../../styles/variables.scss" as var;
-@use "sass:color";
-
-.Field {
-  margin-bottom: 3rem;
-
-  &__label {
-    margin-bottom: 1rem;
-  }
-
-  &__error {
-    color: red;
-    font-size: 0.8em;
-    margin-top: 1rem;
-  }
-
-  &__optional-text {
-    font-size: 0.8em;
-    margin-top: 1rem;
-  }
-
-  input {
-    outline: none;
-    border: solid 1px color.adjust(var.$dark, $alpha: -0.5);
-    font-size: inherit;
-
-    &[type="text"],
-    &[type="email"],
-    &[type="password"],
-    &[type="number"] {
-      width: 100%;
-      padding: 0.8rem 0.6rem;
-      border-radius: 4px;
-    }
-
-    &:focus {
-      background-color: var.$tertiary;
-    }
-  }
-}
