+++ src/components/forms/fields/NumberField.tsx
@@ -1,38 +1,35 @@
 import React from "react";
-import "../Field.scss";
+import "./Field.scss";
 
-export interface ISelectFieldProps {
+export interface INumberFieldProps {
   id: string;
   label: string;
-  options: Array<{
-    label: string;
-    value: any;
-  }>;
-  onChange: React.ChangeEventHandler<HTMLSelectElement>;
   value: string;
+  onChange: React.ChangeEventHandler<HTMLInputElement>;
   required?: boolean;
   error?: string;
 }
 
-export default function SelectField({
+export default function NumberField({
   id,
   label,
-  options,
-  onChange,
   value,
+  onChange,
   required = false,
   error = "",
-}: ISelectFieldProps) {
+}: INumberFieldProps) {
   return (
     <div className="Field">
       <label className="Field__label" htmlFor={id}>
         {label}
       </label>
-      <select id={id} value={value} onChange={onChange}>
-        {options.map(({ label, value }) => {
-          return <option value={value}>{label}</option>;
-        })}
-      </select>
+      <input
+        required={required}
+        type="number"
+        id={id}
+        onChange={onChange}
+        value={value}
+      />
       <div className="Field__optional-text">
         {!required ? "(optional)" : null}
       </div>
