+++ src/components/forms/fields/SelectField.tsx
@@ -1,29 +1,38 @@
 import React from "react";
-import "../Field.scss";
+import "./Field.scss";
 
-export interface ITextAreaFieldProps {
+export interface ISelectFieldProps {
   id: string;
   label: string;
+  options: Array<{
+    label: string;
+    value: any;
+  }>;
+  onChange: React.ChangeEventHandler<HTMLSelectElement>;
   value: string;
-  onChange: React.ChangeEventHandler<HTMLTextAreaElement>;
   required?: boolean;
   error?: string;
 }
 
-export default function TextAreaField({
+export default function SelectField({
   id,
   label,
-  value,
+  options,
   onChange,
+  value,
   required = false,
   error = "",
-}: ITextAreaFieldProps) {
+}: ISelectFieldProps) {
   return (
     <div className="Field">
       <label className="Field__label" htmlFor={id}>
         {label}
       </label>
-      <textarea id={id} onChange={onChange} value={value} required={required} />
+      <select id={id} value={value} onChange={onChange}>
+        {options.map(({ label, value }) => {
+          return <option value={value}>{label}</option>;
+        })}
+      </select>
       <div className="Field__optional-text">
         {!required ? "(optional)" : null}
       </div>
