+++ src/components/forms/fields/index.ts
@@ -2,5 +2,6 @@ import TextField from "./TextField";
 import NumberField from "./NumberField";
 import SelectField from "./SelectField";
 import TextAreaField from "./TextAreaField";
+import PasswordField from "./PasswordField";
 
-export { TextField, NumberField, SelectField, TextAreaField };
+export { TextField, NumberField, SelectField, TextAreaField, PasswordField };
